requirements.txt
m2x==4.0.0
geopy==1.10.0
python-forecastio==1.3.2
APScheduler==3.0.0
